refactor(fleet-console): extract composer building blocks into sdk/composer
Move EffortTrack to the new sdk/composer package (old subpath re-exports
for compatibility) and add ComposerField/Bar/Chip/Input/RestStrip/
AttachControl/SubmitButton plus shared attach/submit icons as pure,
props-driven blocks. Quick Launch reassembles the same surfaces from
these blocks with unchanged class names, DOM, and behavior, so the bar
keeps its exact look including the effort-track apex and ultracode
presentations. This is the component-level single source the upcoming
in-panel chat composer will assemble from.
