+++ runtime/fleet-console/core/client/src/components/composer-blocks.ts
@@ -1,0 +1,13 @@
+// 컴포저 빌딩블록의 core client 셔임 — effort-track.tsx와 같은 갈래다. 컴포넌트는 SDK가
+// 소유하고(플러그인과 같은 부품을 쓴다), core는 경로만 접는다.
+export {
+  AttachImageIcon,
+  ComposerAttachControl,
+  ComposerBar,
+  ComposerChip,
+  ComposerField,
+  ComposerInput,
+  ComposerRestStrip,
+  ComposerSubmitButton,
+  SubmitArrowIcon,
+} from "@fleet-console/sdk/composer";
